// tools/protocol/capnp-v4/generated/current/wire-frame.ts
export const WireFrame_Which = {
  REQUEST: 0,
  RESPONSE: 1,
  SIGNAL: 2,
  /**
* Evolution probe: current decoders reject this variant. The v0 fixture
* does not know its discriminant, which exercises unknown-union handling.
*
*/
  FUTURE_FRAME: 3
} as const;
export type WireFrame_Which = (typeof WireFrame_Which)[keyof typeof WireFrame_Which];
// ...
export class Request extends $.Struct {
  static readonly _capnp = {
    displayName: "Request",
    id: "ef8871a9cce332a2",
    size: new $.ObjectSize(0, 6),
  };
  get id(): string {
    return $.utils.getText(0, this);
  }
  set id(value: string) {
    $.utils.setText(0, value, this);
  }
  get call(): string {
    return $.utils.getText(1, this);
  }
  set call(value: string) {
    $.utils.setText(1, value, this);
  }
  _adoptArgs(value: $.Orphan<JsonValue>): void {
    $.utils.adopt(value, $.utils.getPointer(2, this));
  }
  _disownArgs(): $.Orphan<JsonValue> {
    return $.utils.disown(this.args);
  }
  get args(): JsonValue {
    return $.utils.getStruct(2, JsonValue, this);
  }
  _hasArgs(): boolean {
    return !$.utils.isNull($.utils.getPointer(2, this));
  }
  _initArgs(): JsonValue {
    return $.utils.initStructAt(2, JsonValue, this);
  }
  set args(value: JsonValue) {
    $.utils.copyFrom(value, $.utils.getPointer(2, this));
  }
  get runId(): string {
    return $.utils.getText(3, this);
  }
  set runId(value: string) {
    $.utils.setText(3, value, this);
  }
  _adoptBody(value: $.Orphan<BodyDescriptor>): void {
    $.utils.adopt(value, $.utils.getPointer(4, this));
  }
  _disownBody(): $.Orphan<BodyDescriptor> {
    return $.utils.disown(this.body);
  }
  get body(): BodyDescriptor {
    return $.utils.getStruct(4, BodyDescriptor, this);
  }
  _hasBody(): boolean {
    return !$.utils.isNull($.utils.getPointer(4, this));
  }
  _initBody(): BodyDescriptor {
    return $.utils.initStructAt(4, BodyDescriptor, this);
  }
  set body(value: BodyDescriptor) {
    $.utils.copyFrom(value, $.utils.getPointer(4, this));
  }
  /**
* Evolution probe: ordinary encoders leave this null. The v0 fixture lacks
* the field so tests can verify old-reader/new-writer behavior.
*
*/
  get revisionProbe(): string {
    return $.utils.getText(5, this);
  }
  set revisionProbe(value: string) {
    $.utils.setText(5, value, this);
  }
  toString(): string { return "Request_" + super.toString(); }
}
export const Response_Which = {
  SUCCESS: 0,
  FAILURE: 1
} as const;
export type Response_Which = (typeof Response_Which)[keyof typeof Response_Which];
export class Response extends $.Struct {
  static readonly SUCCESS = Response_Which.SUCCESS;
  static readonly FAILURE = Response_Which.FAILURE;
  static readonly _capnp = {
    displayName: "Response",
    id: "9dc8977a68f3aaba",
    size: new $.ObjectSize(8, 2),
  };
  get id(): string {
    return $.utils.getText(0, this);
  }
  set id(value: string) {
    $.utils.setText(0, value, this);
  }
  _adoptSuccess(value: $.Orphan<ResponseSuccess>): void {
    $.utils.setUint16(0, 0, this);
    $.utils.adopt(value, $.utils.getPointer(1, this));
  }
  _disownSuccess(): $.Orphan<ResponseSuccess> {
    return $.utils.disown(this.success);
  }
  get success(): ResponseSuccess {
    $.utils.testWhich("success", $.utils.getUint16(0, this), 0, this);
    return $.utils.getStruct(1, ResponseSuccess, this);
  }
  _hasSuccess(): boolean {
    return !$.utils.isNull($.utils.getPointer(1, this));
  }
  _initSuccess(): ResponseSuccess {
    $.utils.setUint16(0, 0, this);
    return $.utils.initStructAt(1, ResponseSuccess, this);
  }
  get _isSuccess(): boolean {
    return $.utils.getUint16(0, this) === 0;
  }
  set success(value: ResponseSuccess) {
    $.utils.setUint16(0, 0, this);
    $.utils.copyFrom(value, $.utils.getPointer(1, this));
  }
  _adoptFailure(value: $.Orphan<ResponseFailure>): void {
    $.utils.setUint16(0, 1, this);
    $.utils.adopt(value, $.utils.getPointer(1, this));
  }
  _disownFailure(): $.Orphan<ResponseFailure> {
    return $.utils.disown(this.failure);
  }
  get failure(): ResponseFailure {
    $.utils.testWhich("failure", $.utils.getUint16(0, this), 1, this);
    return $.utils.getStruct(1, ResponseFailure, this);
  }
  _hasFailure(): boolean {
    return !$.utils.isNull($.utils.getPointer(1, this));
  }
  _initFailure(): ResponseFailure {
    $.utils.setUint16(0, 1, this);
    return $.utils.initStructAt(1, ResponseFailure, this);
  }
  get _isFailure(): boolean {
    return $.utils.getUint16(0, this) === 1;
  }
  set failure(value: ResponseFailure) {
    $.utils.setUint16(0, 1, this);
    $.utils.copyFrom(value, $.utils.getPointer(1, this));
  }
  toString(): string { return "Response_" + super.toString(); }
  which(): Response_Which {
    return $.utils.getUint16(0, this) as Response_Which;
  }
}
// ...
export class ResponseFailure extends $.Struct {
  static readonly _capnp = {
    displayName: "ResponseFailure",
    id: "982a6370ca311a13",
    size: new $.ObjectSize(0, 1),
  };
  _adoptError(value: $.Orphan<WireError>): void {
    $.utils.adopt(value, $.utils.getPointer(0, this));
  }
  _disownError(): $.Orphan<WireError> {
    return $.utils.disown(this.error);
  }
  get error(): WireError {
    return $.utils.getStruct(0, WireError, this);
  }
  _hasError(): boolean {
    return !$.utils.isNull($.utils.getPointer(0, this));
  }
  _initError(): WireError {
    return $.utils.initStructAt(0, WireError, this);
  }
  set error(value: WireError) {
    $.utils.copyFrom(value, $.utils.getPointer(0, this));
  }
  toString(): string { return "ResponseFailure_" + super.toString(); }
}
export const Signal_Seq_Which = {
  ABSENT: 0,
  VALUE: 1
} as const;
export type Signal_Seq_Which = (typeof Signal_Seq_Which)[keyof typeof Signal_Seq_Which];
// ...
export const WireError_Retryable_Which = {
  ABSENT: 0,
  VALUE: 1
} as const;
export type WireError_Retryable_Which = (typeof WireError_Retryable_Which)[keyof typeof WireError_Retryable_Which];
// ...
export const BodyDescriptor_Length_Which = {
  ABSENT: 0,
  VALUE: 1
} as const;
export type BodyDescriptor_Length_Which = (typeof BodyDescriptor_Length_Which)[keyof typeof BodyDescriptor_Length_Which];
// ...
export class BodyDescriptor extends $.Struct {
  static readonly _capnp = {
    displayName: "BodyDescriptor",
    id: "fbcff0aefbbaff85",
    size: new $.ObjectSize(16, 0),
  };
  get streamId(): number {
    return $.utils.getUint32(0, this);
  }
  set streamId(value: number) {
    $.utils.setUint32(0, value, this);
  }
  get length(): BodyDescriptor_Length {
    return $.utils.getAs(BodyDescriptor_Length, this);
  }
  _initLength(): BodyDescriptor_Length {
    return $.utils.getAs(BodyDescriptor_Length, this);
  }
  toString(): string { return "BodyDescriptor_" + super.toString(); }
}
export const JsonValue_Which = {
  NULL_VALUE: 0,
  BOOL_VALUE: 1,
  NUMBER_VALUE: 2,
  STRING_VALUE: 3,
  ARRAY_VALUE: 4,
  OBJECT_VALUE: 5
} as const;
export type JsonValue_Which = (typeof JsonValue_Which)[keyof typeof JsonValue_Which];
export class JsonValue extends $.Struct {
  static readonly NULL_VALUE = JsonValue_Which.NULL_VALUE;
  static readonly BOOL_VALUE = JsonValue_Which.BOOL_VALUE;
  static readonly NUMBER_VALUE = JsonValue_Which.NUMBER_VALUE;
  static readonly STRING_VALUE = JsonValue_Which.STRING_VALUE;
  static readonly ARRAY_VALUE = JsonValue_Which.ARRAY_VALUE;
  static readonly OBJECT_VALUE = JsonValue_Which.OBJECT_VALUE;
  static readonly _capnp = {
    displayName: "JsonValue",
    id: "88e2e5ab609c8b47",
    size: new $.ObjectSize(16, 1),
  };
  static _ArrayValue: $.ListCtor<JsonValue>;
  static _ObjectValue: $.ListCtor<JsonEntry>;
  get _isNullValue(): boolean {
    return $.utils.getUint16(0, this) === 0;
  }
  set nullValue(_: true) {
    $.utils.setUint16(0, 0, this);
  }
  get boolValue(): boolean {
    $.utils.testWhich("boolValue", $.utils.getUint16(0, this), 1, this);
    return $.utils.getBit(16, this);
  }
  get _isBoolValue(): boolean {
    return $.utils.getUint16(0, this) === 1;
  }
  set boolValue(value: boolean) {
    $.utils.setUint16(0, 1, this);
    $.utils.setBit(16, value, this);
  }
  get numberValue(): number {
    $.utils.testWhich("numberValue", $.utils.getUint16(0, this), 2, this);
    return $.utils.getFloat64(8, this);
  }
  get _isNumberValue(): boolean {
    return $.utils.getUint16(0, this) === 2;
  }
  set numberValue(value: number) {
    $.utils.setUint16(0, 2, this);
    $.utils.setFloat64(8, value, this);
  }
  get stringValue(): string {
    $.utils.testWhich("stringValue", $.utils.getUint16(0, this), 3, this);
    return $.utils.getText(0, this);
  }
  get _isStringValue(): boolean {
    return $.utils.getUint16(0, this) === 3;
  }
  set stringValue(value: string) {
    $.utils.setUint16(0, 3, this);
    $.utils.setText(0, value, this);
  }
  _adoptArrayValue(value: $.Orphan<$.List<JsonValue>>): void {
    $.utils.setUint16(0, 4, this);
    $.utils.adopt(value, $.utils.getPointer(0, this));
  }
  _disownArrayValue(): $.Orphan<$.List<JsonValue>> {
    return $.utils.disown(this.arrayValue);
  }
  get arrayValue(): $.List<JsonValue> {
    $.utils.testWhich("arrayValue", $.utils.getUint16(0, this), 4, this);
    return $.utils.getList(0, JsonValue._ArrayValue, this);
  }
  _hasArrayValue(): boolean {
    return !$.utils.isNull($.utils.getPointer(0, this));
  }
  _initArrayValue(length: number): $.List<JsonValue> {
    $.utils.setUint16(0, 4, this);
    return $.utils.initList(0, JsonValue._ArrayValue, length, this);
  }
  get _isArrayValue(): boolean {
    return $.utils.getUint16(0, this) === 4;
  }
  set arrayValue(value: $.List<JsonValue>) {
    $.utils.setUint16(0, 4, this);
    $.utils.copyFrom(value, $.utils.getPointer(0, this));
  }
  _adoptObjectValue(value: $.Orphan<$.List<JsonEntry>>): void {
    $.utils.setUint16(0, 5, this);
    $.utils.adopt(value, $.utils.getPointer(0, this));
  }
  _disownObjectValue(): $.Orphan<$.List<JsonEntry>> {
    return $.utils.disown(this.objectValue);
  }
  get objectValue(): $.List<JsonEntry> {
    $.utils.testWhich("objectValue", $.utils.getUint16(0, this), 5, this);
    return $.utils.getList(0, JsonValue._ObjectValue, this);
  }
  _hasObjectValue(): boolean {
    return !$.utils.isNull($.utils.getPointer(0, this));
  }
  _initObjectValue(length: number): $.List<JsonEntry> {
    $.utils.setUint16(0, 5, this);
    return $.utils.initList(0, JsonValue._ObjectValue, length, this);
  }
  get _isObjectValue(): boolean {
    return $.utils.getUint16(0, this) === 5;
  }
  set objectValue(value: $.List<JsonEntry>) {
    $.utils.setUint16(0, 5, this);
    $.utils.copyFrom(value, $.utils.getPointer(0, this));
  }
  toString(): string { return "JsonValue_" + super.toString(); }
  which(): JsonValue_Which {
    return $.utils.getUint16(0, this) as JsonValue_Which;
  }
}
export class JsonEntry extends $.Struct {
  static readonly _capnp = {
    displayName: "JsonEntry",
    id: "e1397dabc978d5ee",
    size: new $.ObjectSize(0, 2),
  };
  get key(): string {
    return $.utils.getText(0, this);
  }
  set key(value: string) {
    $.utils.setText(0, value, this);
  }
  _adoptValue(value: $.Orphan<JsonValue>): void {
    $.utils.adopt(value, $.utils.getPointer(1, this));
  }
  _disownValue(): $.Orphan<JsonValue> {
    return $.utils.disown(this.value);
  }
  get value(): JsonValue {
    return $.utils.getStruct(1, JsonValue, this);
  }
  _hasValue(): boolean {
    return !$.utils.isNull($.utils.getPointer(1, this));
  }
  _initValue(): JsonValue {
    return $.utils.initStructAt(1, JsonValue, this);
  }
  set value(value: JsonValue) {
    $.utils.copyFrom(value, $.utils.getPointer(1, this));
  }
  toString(): string { return "JsonEntry_" + super.toString(); }
}